Reduce config.rb to the settings manifest
ConfigRegistry base class holds the machinery, validators, and a
generic check DSL; the imgproxy all-or-nothing rule becomes a
declared check. Both files move to lib/boot, so zeitwerk and
SimpleCov each exclude one directory instead of listing files.
