+++ test/lib/boot/config_test.rb
